einiges editiert

+++ LenzDefenz/src/helpers/Clock.java
@@ -41,7 +41,7 @@ public class Clock {
 		totalTime += d;
 	}
 	
-	public static void changeMultiplier(int change){
+	public static void changeMultiplier(float change){
 		if (multiplier + change < -1 && multiplier +change >7 ){
 			
 		}else{
